requirements.txt
numpy==1.24.1
Flask
gunicorn
sklearn
pandas==1.5.3
PyYAML
-e .
